Move cartopy.tests.show to cartopy.tests.mpl.
This seems like a more fitting place for it, though it's not really used
for anything, actually.
