+++ lib/cartopy/tests/test_linear_ring.py
@@ -34,7 +34,7 @@ class TestBoundary(unittest.TestCase):
                                                    (10, 50)])
         projection = ccrs.Robinson(170.5)
         multi_line_string = projection.project_geometry(linear_ring)
-        from cartopy.tests import show
+        # from cartopy.tests.mpl import show
         # show(projection, multi_line_string)
 
         # The original ring should have been split into multiple pieces.
@@ -97,7 +97,7 @@ class TestMisc(unittest.TestCase):
                                                    (10, 60),
                                                    (10, 50)])
         multi_line_string = projection.project_geometry(linear_ring)
-        # from cartopy.tests import show
+        # from cartopy.tests.mpl import show
         # show(projection, multi_line_string)
         # XXX not a test...
 
@@ -114,7 +114,7 @@ class TestMisc(unittest.TestCase):
         # there should be one, and only one, returned line:
         assert isinstance(multi_line_string, geometry.polygon.LinearRing)
 
-        # from cartopy.tests import show
+        # from cartopy.tests.mpl import show
         # show(projection, multi_line_string)
 
     def test_three_points(self):
