+++ lib/cartopy/tests/test_polygon.py
@@ -162,7 +162,7 @@ class TestQuality(unittest.TestCase):
             (175.0, -57.19913331),
         ])
         self.multi_polygon = projection.project_geometry(polygon)
-        # from cartopy.tests import show
+        # from cartopy.tests.mpl import show
         # show(projection, self.multi_polygon)
 
     def test_split(self):
